switched to using django-bootstrap3
Made this switch so that form generation would require less (of my)
code. I no longer have to subclass the django-allauth default forms,
and save a lot of text in my template <form>s. It costs a slight amount
of flexibility, but should at least be serviceable for the allauth
forms (I won’t have to subclass my own forms!)